[fix]退出失效等问题处理

+++ DoodieBird/libs/cocos2d/CCDirector.m
@@ -583,7 +583,7 @@ static CCDirector *_sharedDirector = nil;
 			[FPSLabel_ setString:fpsstr];
 			[fpsstr release];
 			
-			NSString *draws = [[NSString alloc] initWithFormat:@"%4d", __ccNumberOfDraws];
+			NSString *draws = [[NSString alloc] initWithFormat:@"%4lu", (unsigned long)__ccNumberOfDraws];
 			[drawsLabel_ setString:draws];
 			[draws release];
 		}
@@ -636,4 +636,3 @@ static CCDirector *_sharedDirector = nil;
 }
 
 @end
-
